requirements.txt
fastapi
uvicorn
transformers
accelerate
bitsandbytes
python-dotenv
torch
pydantic